format, moved some exception definitions
