+++ tableauserverclient/helpers/logging.py
@@ -1,4 +1,3 @@
-
 import logging
 
 logger = logging.getLogger("TSC")
