+++ tableauserverclient/server/endpoint/resource_tagger.py
@@ -1,9 +1,9 @@
 import copy
-import logging
 import urllib.parse
 
 from .endpoint import Endpoint
-from .exceptions import EndpointUnavailableError, ServerResponseError
+from .exceptions import ServerResponseError
+from ..exceptions import EndpointUnavailableError
 from tableauserverclient.server import RequestFactory
 from tableauserverclient.models import TagItem
