+++ tableauserverclient/server/endpoint/schedules_endpoint.py
@@ -10,6 +10,7 @@ from tableauserverclient.server import RequestFactory
 from tableauserverclient.models import PaginationItem, ScheduleItem, TaskItem
 
 from tableauserverclient.helpers.logging import logger
+
 AddResponse = namedtuple("AddResponse", ("result", "error", "warnings", "task_created"))
 OK = AddResponse(result=True, error=None, warnings=None, task_created=None)
