+++ tableauserverclient/server/endpoint/workbooks_endpoint.py
@@ -45,6 +45,7 @@ FILESIZE_LIMIT = 1024 * 1024 * 64  # 64MB
 ALLOWED_FILE_EXTENSIONS = ["twb", "twbx"]
 
 from tableauserverclient.helpers.logging import logger
+
 FilePath = Union[str, os.PathLike]
 FileObject = Union[io.BufferedReader, io.BytesIO]
 FileObjectR = Union[io.BufferedReader, io.BytesIO]
