+++ tableauserverclient/server/exceptions.py
@@ -1,2 +1,8 @@
-class NotSignedInError(Exception):
+# These errors can be thrown without even talking to Tableau Server
+
+class ServerInfoEndpointNotFoundError(Exception):
+    pass
+
+
+class EndpointUnavailableError(Exception):
     pass
